# Optimize find_nearest_stations query performance