Summary:
This commit significantly improves the performance of the find_nearest_stations function. Changes include applying filters directly in the database query to reduce the dataset early in the process and optimizing loop efficiency to handle distance calculations and grouping more effectively. These optimizations have resulted in a notable decrease in execution time, especially when filters are applied.

Details:
- Implemented early filtering in the database query for `service_providers` and `frequency_bands`, reducing the number of stations processed in subsequent steps
- Combined the logic of filtering by `max_distance` and grouping stations by unique location into a single loop to avoid multiple iterations over the stations
- Adjusted the sorting and limiting of stations to apply after grouping, ensuring correct application of the `limit` parameter

Performance Improvements:
- Without filters, the execution time improved by approximately 1 second, decrease from 4 seconds to 3 seconds
- With filters applied, the execution time dropped from 3 seconds to 0.5 seconds, demonstrating a significant improvement in query efficiency

These changes enhance the function's efficiency, providing faster responses to user queries and reducing server load, especially under high-demand scenarios.